Extract C++ integration code from required toolchain
This refactors the crates so that we have:
  * ecap: core crate, defines traits and structs (similar to libecap itself)
  * ecap-common: shared library which provides service/translator registration
  * ecap-common-link: workaround for Cargo, shim over ecap-common so
    that crates don't need build scripts
  * ecap-cpp: translator from C++ to Rust types (currently incomplete)

Note that this will eventually permit writing Rust services which
integrate with Rust hosts, without going through C/C++.

The model is that a C++ host which supports libecap will load the
translator and service modules separately, and after registration
through ecap-common, they will begin talking directly to each other
through the trait-based APIs defined in `ecap`.

The currently implemented model is not optimal: we allocate services and
translators multiple times, despite talking to Rust. However this
doesn't matter too much since registration occurs once, and the
overheads are not excessive.